restructured folders and added eval pipeline
