+++ recipes/use_cases/end2end-recipes/chatbot/pipelines/config.py
