+++ recipes/use_cases/end2end-recipes/chatbot/pipelines/doc_processor.py
